+++ backend/src/main/java/moim_today/dto/moim/moim/MoimDeleteRequest.java
@@ -1,4 +1,4 @@
-package moim_today.dto.moim;
+package moim_today.dto.moim.moim;
 
 public record MoimDeleteRequest(
         long moimId
