+++ backend/src/test/java/moim_today/fake_class/moim/FakeMoimService.java
@@ -1,12 +1,12 @@
 package moim_today.fake_class.moim;
 
-import moim_today.application.moim.MoimService;
+import moim_today.application.moim.moim.MoimService;
 import moim_today.domain.moim.DisplayStatus;
 import moim_today.domain.moim.enums.MoimCategory;
-import moim_today.dto.moim.MoimAppendRequest;
-import moim_today.dto.moim.MoimDetailResponse;
-import moim_today.dto.moim.MoimUpdateRequest;
-import moim_today.dto.moim.MoimImageResponse;
+import moim_today.dto.moim.moim.MoimAppendRequest;
+import moim_today.dto.moim.moim.MoimDetailResponse;
+import moim_today.dto.moim.moim.MoimUpdateRequest;
+import moim_today.dto.moim.moim.MoimImageResponse;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDate;
