+++ backend/src/test/java/moim_today/presentation/moim/MoimControllerTest.java
@@ -3,9 +3,9 @@ package moim_today.presentation.moim;
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import moim_today.domain.moim.DisplayStatus;
 import moim_today.domain.moim.enums.MoimCategory;
-import moim_today.dto.moim.MoimAppendRequest;
-import moim_today.dto.moim.MoimDeleteRequest;
-import moim_today.dto.moim.MoimUpdateRequest;
+import moim_today.dto.moim.moim.MoimAppendRequest;
+import moim_today.dto.moim.moim.MoimDeleteRequest;
+import moim_today.dto.moim.moim.MoimUpdateRequest;
 import moim_today.fake_class.moim.FakeMoimService;
 import moim_today.util.ControllerTest;
 import org.junit.jupiter.api.DisplayName;
